requirements.txt
beautifulsoup4 (4.5.1)
bs4 (0.0.1)
cffi (1.9.1)
constantly (15.1.0)
cssselect (1.0.0)
cycler (0.10.0)
idna (2.2)
incremental (16.10.1)
lxml (3.7.1)
matplotlib (1.5.3)
mpmath (0.19)
nltk (3.2.1)
numpy (1.11.2)
oauthlib (2.0.1)
pandas (0.19.1)
parsel (1.1.0)
Pillow (3.4.2)
pip (9.0.1)
pyasn1 (0.1.9)
pycparser (2.17)
PyDispatcher (2.0.5)
pyparsing (2.1.10)
python-dateutil (2.6.0)
pytz (2016.7)
requests (2.12.4)
requests-oauthlib (0.7.0)
scikit-learn (0.18.1)
scipy (0.18.1)
setuptools (29.0.1)
six (1.10.0)
sympy (1.0)
textblob (0.11.1)
tweepy (3.5.0)
w3lib (1.16.0)
wheel (0.29.0)
zope.interface (4.3.3)
